separate functions from api
